+++ pacman-core/src/main/java/de/amr/games/pacman/model/pacman/PacManGame.java
@@ -121,14 +121,14 @@ public class PacManGame extends GameModel {
 		checkLevelNumber(levelNumber);
 		//@formatter:off
 		var bonus = switch (levelNumber) {
-		case 1      -> new StaticBonus(0,  100); // Cherries
-		case 2      -> new StaticBonus(1,  300); // Strawberry
-		case 3, 4   -> new StaticBonus(2,  500); // Peach
-		case 5, 6   -> new StaticBonus(3,  700); // Apple
-		case 7, 8   -> new StaticBonus(4, 1000); // Grapes
-		case 9, 10  -> new StaticBonus(5, 2000); // Galaxian
-		case 11, 12 -> new StaticBonus(6, 3000); // Bell
-		default     -> new StaticBonus(7, 5000); // Key
+		case 1      -> new StaticBonus((byte)0,  100); // Cherries
+		case 2      -> new StaticBonus((byte)1,  300); // Strawberry
+		case 3, 4   -> new StaticBonus((byte)2,  500); // Peach
+		case 5, 6   -> new StaticBonus((byte)3,  700); // Apple
+		case 7, 8   -> new StaticBonus((byte)4, 1000); // Grapes
+		case 9, 10  -> new StaticBonus((byte)5, 2000); // Galaxian
+		case 11, 12 -> new StaticBonus((byte)6, 3000); // Bell
+		default     -> new StaticBonus((byte)7, 5000); // Key
 		};
 		//@formatter:on
 		bonus.entity().setPosition(halfTileRightOf(v2i(13, 20)));
